Get titles functionality moved to catalog resource
