+++ graylog2-server/src/main/java/org/graylog2/rest/resources/system/contentpacks/CatalogResource.java
@@ -31,11 +31,16 @@ import org.graylog2.contentpacks.model.entities.EntityExcerpt;
 import org.graylog2.rest.models.system.contentpacks.responses.CatalogIndexResponse;
 import org.graylog2.rest.models.system.contentpacks.responses.CatalogResolveRequest;
 import org.graylog2.rest.models.system.contentpacks.responses.CatalogResolveResponse;
+import org.graylog2.rest.resources.system.contentpacks.titles.EntityTitleService;
+import org.graylog2.rest.resources.system.contentpacks.titles.model.EntitiesTitleResponse;
+import org.graylog2.rest.resources.system.contentpacks.titles.model.EntityTitleRequest;
+import org.graylog2.shared.rest.resources.RestResource;
 import org.graylog2.shared.security.RestPermissions;
 
 import javax.inject.Inject;
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
+import javax.ws.rs.Consumes;
 import javax.ws.rs.GET;
 import javax.ws.rs.POST;
 import javax.ws.rs.Path;
@@ -47,12 +52,15 @@ import java.util.Set;
 @Api(value = "System/Catalog", description = "Entity Catalog")
 @Path("/system/catalog")
 @Produces(MediaType.APPLICATION_JSON)
-public class CatalogResource {
+public class CatalogResource extends RestResource {
     private final ContentPackService contentPackService;
+    private final EntityTitleService entityTitleService;
 
     @Inject
-    public CatalogResource(ContentPackService contentPackService) {
+    public CatalogResource(final ContentPackService contentPackService,
+                           final EntityTitleService entityTitleService) {
         this.contentPackService = contentPackService;
+        this.entityTitleService = entityTitleService;
     }
 
     @GET
@@ -77,5 +85,15 @@ public class CatalogResource {
         final ImmutableSet<Entity> entities = contentPackService.collectEntities(resolvedEntities);
 
         return CatalogResolveResponse.create(entities);
+    }
+
+    @POST
+    @Timed
+    @ApiOperation(value = "Get titles of provided entities")
+    @NoAuditEvent("This endpoint does not change any data")
+    @Consumes(MediaType.APPLICATION_JSON)
+    @Path("/entity_titles")
+    public EntitiesTitleResponse getTitles(@ApiParam(name = "JSON body", required = true) final EntityTitleRequest request) {
+        return entityTitleService.getTitles(getSubject(), request);
     }
 }
