+++ graylog2-server/src/main/java/org/graylog2/rest/resources/system/contentpacks/titles/model/EntityIdentifier.java
@@ -14,11 +14,10 @@
  * along with this program. If not, see
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
-package org.graylog2.rest.resources.entities.titles.model;
+package org.graylog2.rest.resources.system.contentpacks.titles.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public record EntityTitleResponse(@JsonProperty("id") String id,
-                                  @JsonProperty("type") String type,
-                                  @JsonProperty("title") String title) {
+public record EntityIdentifier(@JsonProperty("id") String id,
+                               @JsonProperty("type") String type) {
 }
