+++ graylog2-server/src/main/java/org/graylog2/rest/resources/system/contentpacks/titles/model/EntityTitleResponse.java
@@ -14,10 +14,11 @@
  * along with this program. If not, see
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
-package org.graylog2.rest.resources.entities.titles.model;
+package org.graylog2.rest.resources.system.contentpacks.titles.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public record EntityIdentifier(@JsonProperty("id") String id,
-                               @JsonProperty("type") String type) {
+public record EntityTitleResponse(@JsonProperty("id") String id,
+                                  @JsonProperty("type") String type,
+                                  @JsonProperty("title") String title) {
 }
