+++ graylog2-server/src/test/java/org/graylog2/rest/resources/system/contentpacks/titles/EntityTitleServiceTest.java
@@ -14,17 +14,17 @@
  * along with this program. If not, see
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
-package org.graylog2.rest.resources.entities.titles;
+package org.graylog2.rest.resources.system.contentpacks.titles;
 
 import org.apache.shiro.subject.Subject;
 import org.graylog2.database.DbEntity;
 import org.graylog2.database.MongoConnection;
 import org.graylog2.database.dbcatalog.DbEntitiesCatalog;
 import org.graylog2.database.dbcatalog.DbEntityCatalogEntry;
-import org.graylog2.rest.resources.entities.titles.model.EntitiesTitleResponse;
-import org.graylog2.rest.resources.entities.titles.model.EntityIdentifier;
-import org.graylog2.rest.resources.entities.titles.model.EntityTitleRequest;
-import org.graylog2.rest.resources.entities.titles.model.EntityTitleResponse;
+import org.graylog2.rest.resources.system.contentpacks.titles.model.EntitiesTitleResponse;
+import org.graylog2.rest.resources.system.contentpacks.titles.model.EntityIdentifier;
+import org.graylog2.rest.resources.system.contentpacks.titles.model.EntityTitleRequest;
+import org.graylog2.rest.resources.system.contentpacks.titles.model.EntityTitleResponse;
 import org.graylog2.streams.StreamImpl;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -36,7 +36,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 
-import static org.graylog2.rest.resources.entities.titles.model.EntitiesTitleResponse.EMPTY_RESPONSE;
+import static org.graylog2.rest.resources.system.contentpacks.titles.model.EntitiesTitleResponse.EMPTY_RESPONSE;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertSame;
 import static org.mockito.Mockito.doReturn;
